+++ bizdemo/tiktok_demo/biz/dal/db/follows.go
@@ -26,7 +26,7 @@ import (
 	"github.com/cloudwego/hertz-examples/bizdemo/tiktok_demo/pkg/constants"
 )
 
-// Follows user_id 关注了 follower_id
+// Follows user_id follow follower_id
 type Follows struct {
 	ID         int64          `json:"id"`
 	UserId     int64          `json:"user_id"`
@@ -35,6 +35,7 @@ type Follows struct {
 	DeletedAt  gorm.DeletedAt `gorm:"index" json:"delete_at"`
 }
 
+// TableName set table name to make gorm can correctly identify
 func (Follows) TableName() string {
 	return constants.FollowsTableName
 }
@@ -44,7 +45,7 @@ func AddNewFollow(follow *Follows) (bool, error) {
 	if err != nil {
 		return false, err
 	}
-	// 如果缓存中有此关系，更新
+	// if redis hit update
 	if cnt, _ := redis.RdbFollowing.SCard(strconv.Itoa(int(follow.UserId))).Result(); cnt > 0 {
 		redis.RdbFollowing.SAdd(strconv.Itoa(int(follow.UserId)), follow.FollowerId)
 		redis.RdbFollower.Expire(strconv.Itoa(int(follow.UserId)), redis.ExpireTime)
@@ -56,12 +57,13 @@ func AddNewFollow(follow *Follows) (bool, error) {
 	return true, nil
 }
 
+// DeleteFollow delete follow relation in db and update redis
 func DeleteFollow(follow *Follows) (bool, error) {
 	err := DB.Where("user_id = ? AND follower_id = ?", follow.UserId, follow.FollowerId).Delete(follow).Error
 	if err != nil {
 		return false, err
 	}
-	// 如果缓存中有此关系，更新
+	// if redis hit update
 	if cnt, _ := redis.RdbFollowing.SCard(strconv.Itoa(int(follow.UserId))).Result(); cnt > 0 {
 		redis.RdbFollowing.SRem(strconv.Itoa(int(follow.UserId)), follow.FollowerId)
 		redis.RdbFollower.Expire(strconv.Itoa(int(follow.UserId)), redis.ExpireTime)
@@ -73,81 +75,79 @@ func DeleteFollow(follow *Follows) (bool, error) {
 	return true, nil
 }
 
-func QueryFollowExist(follow *Follows) (bool, error) {
-	if exist, err := redis.RdbFollowing.SIsMember(strconv.Itoa(int(follow.UserId)), follow.FollowerId).Result(); exist {
-		// fmt.Printf("在redis中获取到user_id:%d关注了user:%d的关系\n", follow.UserId, follow.FollowerId)
-		redis.RdbFollowing.Expire(strconv.Itoa(int(follow.UserId)), redis.ExpireTime)
+// QueryFollowExist check the relation of user and follower
+func QueryFollowExist(user_id, follower_id int64) (bool, error) {
+	if exist, err := redis.RdbFollowing.SIsMember(strconv.Itoa(int(user_id)), follower_id).Result(); exist {
+		redis.RdbFollowing.Expire(strconv.Itoa(int(user_id)), redis.ExpireTime)
 		return true, err
 	}
-	err := DB.Where("user_id = ? AND follower_id = ?", follow.UserId, follow.FollowerId).Find(&follow).Error
+	follow := Follows{
+		UserId:     user_id,
+		FollowerId: follower_id,
+	}
+	err := DB.Where("user_id = ? AND follower_id = ?", user_id, follower_id).Find(&follow).Error
 	if err != nil {
 		return false, err
 	}
 	if follow.ID == 0 {
 		return false, nil
 	}
-	// 异步更新redis
+	// Update redis asynchronously
 	go addRelationToRedis(int(follow.UserId), int(follow.FollowerId))
 	return true, nil
 }
 
-// GetFollowCount 查询用户的关注数量
+// GetFollowCount query the number of users following
 func GetFollowCount(user_id int64) (int64, error) {
 	if count, err := redis.RdbFollowing.SCard(strconv.Itoa(int(user_id))).Result(); count > 0 {
-		// 更新过期时间。
+		// update expiration time
 		redis.RdbFollowing.Expire(strconv.Itoa(int(user_id)), redis.ExpireTime)
-		// fmt.Printf("在redis中获取到user_id:%d的关注数%d\n", user_id, count)
 		return count, err
 	}
-	// 缓存中没有，去数据库查找并更新缓存
+	// Not in the cache, go to the database to find and update the cache
 	followings, err := GetFollowIdList(user_id)
 	if err != nil {
 		return 0, err
 	}
-	// 更新Redis
+	// update cache
 	go AddNewFollowRelationToRedis(user_id, followings)
 	return int64(len(followings)), nil
 }
 
-// AddNewFollowRelationToRedis 更新redis.RdbFollowing
-
+// AddNewFollowRelationToRedis update redis.RdbFollowing
 func AddNewFollowRelationToRedis(user_id int64, followings []int64) {
 	for _, following := range followings {
-		// redis.RdbFollowing.SAdd(strconv.Itoa(int(user_id)), -1)
 		redis.RdbFollowing.SAdd(strconv.Itoa(int(user_id)), following)
 	}
-	// 更新过期时间，保持数据热度
+	// update expiration time to keep data hot
 	redis.RdbFollowing.Expire(strconv.Itoa(int(user_id)), redis.ExpireTime)
 }
 
-// GetFolloweeCount 查询用户的粉丝数量
+// GetFolloweeCount query the number of followers of a user
 func GetFolloweeCount(follower_id int64) (int64, error) {
 	if count, err := redis.RdbFollower.SCard(strconv.Itoa(int(follower_id))).Result(); count > 0 {
-		// 更新过期时间。
+		// update expire time
 		redis.RdbFollower.Expire(strconv.Itoa(int(follower_id)), redis.ExpireTime)
-		// fmt.Printf("在redis中获取到user_id:%d的粉丝数%d\n", follower_id, count)
 		return count, err
 	}
-	// 缓存中没有，去数据库查找并更新缓存
+	// Not in the cache, go to the database to find and update the cache
 	followers, err := GetFollowerIdList(follower_id)
 	if err != nil {
 		return 0, err
 	}
-	// 更新Redis
 	go AddNewFollowerRelationToRedis(follower_id, followers)
 	return int64(len(followers)), nil
 }
 
-// AddNewFollowerRelationToRedis 更新redis.RdbFollower
+// AddNewFollowerRelationToRedis update redis.RdbFollower
 func AddNewFollowerRelationToRedis(user_id int64, followers []int64) {
 	for _, follower := range followers {
 		redis.RdbFollower.SAdd(strconv.Itoa(int(user_id)), follower)
 	}
-	// 更新过期时间，保持数据热度
 	redis.RdbFollower.Expire(strconv.Itoa(int(user_id)), redis.ExpireTime)
 }
 
-// GetFollowIdList 获得 user_id 关注的人的 id
+// GetFollowIdList find user_id follow id list
 func GetFollowIdList(user_id int64) ([]int64, error) {
 	var follow_actions []Follows
 	err := DB.Where("user_id = ?", user_id).Find(&follow_actions).Error
@@ -161,7 +161,7 @@ func GetFollowIdList(user_id int64) ([]int64, error) {
 	return result, nil
 }
 
-// GetFollowerIdList 获得 follower_id 所有粉丝的 id
+// GetFollowerIdList get follower id list
 func GetFollowerIdList(follower_id int64) ([]int64, error) {
 	var follow_actions []Follows
 	err := DB.Where("follower_id = ?", follower_id).Find(&follow_actions).Error
@@ -175,9 +175,9 @@ func GetFollowerIdList(follower_id int64) ([]int64, error) {
 	return result, nil
 }
 
+// CheckFollowRelationExist use user id and follower id to check the relation
 func CheckFollowRelationExist(follow *Follows) (bool, error) {
 	if exist, err := redis.RdbFollowing.SIsMember(strconv.Itoa(int(follow.UserId)), follow.FollowerId).Result(); exist {
-		// fmt.Printf("在redis中获取到user_id:%d关注了user:%d的关系\n", follow.UserId, follow.FollowerId)
 		redis.RdbFollowing.Expire(strconv.Itoa(int(follow.UserId)), redis.ExpireTime)
 		return true, err
 	}
@@ -188,11 +188,11 @@ func CheckFollowRelationExist(follow *Follows) (bool, error) {
 	if follow.ID == 0 {
 		return false, nil
 	}
-	// 异步更新redis
 	go addRelationToRedis(int(follow.UserId), int(follow.FollowerId))
 	return true, nil
 }
 
+// addRelationToRedis add relation and extend expiration time
 func addRelationToRedis(user_id, follow_id int) {
 	redis.RdbFollowing.SAdd(strconv.Itoa(user_id), follow_id)
 	redis.RdbFollowing.Expire(strconv.Itoa(user_id), redis.ExpireTime)
