+++ bizdemo/tiktok_demo/biz/dal/db/friend_test.go
@@ -29,7 +29,7 @@ func TestGetFriendList(t *testing.T) {
 		return
 	}
 	for _, followerId := range followerList {
-		isFriend, err := QueryFollowExist(&Follows{UserId: 1001, FollowerId: followerId})
+		isFriend, err := QueryFollowExist(1001, followerId)
 		if err != nil {
 			fmt.Println("false")
 			return
