+++ bizdemo/tiktok_demo/biz/dal/db/user.go
@@ -43,7 +43,7 @@ func CreateUser(user *User) (int64, error) {
 	return user.ID, err
 }
 
-// 根据 user_name 查询 User
+// QueryUser query User by user_name
 func QueryUser(userName string) (*User, error) {
 	var user User
 	if err := DB.Where("user_name = ?", userName).Find(&user).Error; err != nil {
@@ -52,6 +52,7 @@ func QueryUser(userName string) (*User, error) {
 	return &user, nil
 }
 
+// QueryUserById get user in the database by user id
 func QueryUserById(user_id int64) (*User, error) {
 	var user User
 	if err := DB.Where("id = ?", user_id).Find(&user).Error; err != nil {
@@ -64,6 +65,7 @@ func QueryUserById(user_id int64) (*User, error) {
 	return &user, nil
 }
 
+// VerifyUser verify username and password in the db
 func VerifyUser(userName, password string) (int64, error) {
 	var user User
 	if err := DB.Where("user_name = ? AND password = ?", userName, password).Find(&user).Error; err != nil {
@@ -76,6 +78,7 @@ func VerifyUser(userName, password string) (int64, error) {
 	return user.ID, nil
 }
 
+// CheckUserExistById find if user exists
 func CheckUserExistById(user_id int64) (bool, error) {
 	var user User
 	if err := DB.Where("id = ?", user_id).Find(&user).Error; err != nil {
