+++ bizdemo/tiktok_demo/biz/handler/ping.go
@@ -26,7 +26,7 @@ import (
 	"github.com/cloudwego/hertz/pkg/protocol/consts"
 )
 
-// Ping .
+// Ping return pong to test network
 func Ping(ctx context.Context, c *app.RequestContext) {
 	c.JSON(consts.StatusOK, utils.H{
 		"message": "pong",
