+++ bizdemo/tiktok_demo/biz/mw/ffmpeg/ffmpeg.go
@@ -22,6 +22,7 @@ import (
 	ffmpeg "github.com/u2takey/ffmpeg-go"
 )
 
+// GetSnapshot get the first frame of a video via ffmpeg
 func GetSnapshot(videoPath string) (buf *bytes.Buffer, err error) {
 	buf = bytes.NewBuffer(nil)
 	err = ffmpeg.Input(videoPath).
