+++ bizdemo/tiktok_demo/biz/mw/jwt/jwt.go
@@ -43,6 +43,7 @@ func Init() {
 		TokenLookup: "query:token,form:token",
 		Timeout:     24 * time.Hour,
 		IdentityKey: identity,
+		// Verify password at login
 		Authenticator: func(ctx context.Context, c *app.RequestContext) (interface{}, error) {
 			var loginRequest user.DouyinUserLoginRequest
 			if err := c.BindAndValidate(&loginRequest); err != nil {
@@ -59,6 +60,7 @@ func Init() {
 			c.Set("user_id", user.ID)
 			return user.ID, nil
 		},
+		// Set the payload in the token
 		PayloadFunc: func(data interface{}) jwt.MapClaims {
 			if v, ok := data.(int64); ok {
 				return jwt.MapClaims{
@@ -67,10 +69,12 @@ func Init() {
 			}
 			return jwt.MapClaims{}
 		},
+		// build login response if verify password successfully
 		LoginResponse: func(ctx context.Context, c *app.RequestContext, code int, token string, expire time.Time) {
 			hlog.CtxInfof(ctx, "Login success ，token is issued clientIP: "+c.ClientIP())
 			c.Set("token", token)
 		},
+		// Verify token and get the id of logged-in user
 		Authorizator: func(data interface{}, ctx context.Context, c *app.RequestContext) bool {
 			if v, ok := data.(float64); ok {
 				current_user_id := int64(v)
@@ -80,6 +84,7 @@ func Init() {
 			}
 			return false
 		},
+		// Validation failed, build the message
 		Unauthorized: func(ctx context.Context, c *app.RequestContext, code int, message string) {
 			c.JSON(consts.StatusOK, user.DouyinUserLoginResponse{
 				StatusCode: errno.AuthorizationFailedErrCode,
