+++ bizdemo/tiktok_demo/biz/mw/minio/minio.go
@@ -36,6 +36,7 @@ var (
 	err    error
 )
 
+// MakeBucket create a bucket with a specified name
 func MakeBucket(ctx context.Context, bucketName string) {
 	exists, err := Client.BucketExists(ctx, bucketName)
 	if err != nil {
@@ -52,6 +53,7 @@ func MakeBucket(ctx context.Context, bucketName string) {
 	}
 }
 
+// PutToBucket put the file into the bucket by *multipart.FileHeader
 func PutToBucket(ctx context.Context, bucketName string, file *multipart.FileHeader) (info minio.UploadInfo, err error) {
 	fileObj, _ := file.Open()
 	info, err = Client.PutObject(ctx, bucketName, file.Filename, fileObj, file.Size, minio.PutObjectOptions{})
@@ -59,6 +61,7 @@ func PutToBucket(ctx context.Context, bucketName string, file *multipart.FileHea
 	return info, err
 }
 
+// GetObjURL get the original link of the file in minio
 func GetObjURL(ctx context.Context, bucketName, filename string) (u *url.URL, err error) {
 	exp := time.Hour * 24
 	reqParams := make(url.Values)
@@ -66,11 +69,13 @@ func GetObjURL(ctx context.Context, bucketName, filename string) (u *url.URL, er
 	return u, err
 }
 
+// PutToBucketByBuf put the file into the bucket by *bytes.Buffer
 func PutToBucketByBuf(ctx context.Context, bucketName, filename string, buf *bytes.Buffer) (info minio.UploadInfo, err error) {
 	info, err = Client.PutObject(ctx, bucketName, filename, buf, int64(buf.Len()), minio.PutObjectOptions{})
 	return info, err
 }
 
+// PutToBucketByFilePath put the file into the bucket by filepath
 func PutToBucketByFilePath(ctx context.Context, bucketName, filename, filepath string) (info minio.UploadInfo, err error) {
 	info, err = Client.FPutObject(ctx, bucketName, filename, filepath, minio.PutObjectOptions{})
 	return info, err
